[81] Show count of characters in .pinfo command. Improve .gobject info to not use SQL query anymore. Thanks to silverice for suggesting. Add support for sending quest completion text Add creature expansion checkIf stealth-only spell fails, shouldn't remove you from stealth. Fix by lol584. Closes issue #99. Move initialization of m_GMticketid to avoid random numbers being assigned if gm ticket table is empty. Thanks Ayane. Fix command .go ticket. Will properly work now rather then show syntax error. Fixes issue #465.

+++ src/game/Creature.cpp
@@ -132,7 +132,7 @@ bool AssistDelayEvent::Execute(uint64 /*e_time*/, uint32 /*p_time*/)
     return true;
 }
 
-CreatureBaseStats const* CreatureBaseStats::GetBaseStats(uint32 level, uint8 unitClass)
+CreatureBaseStats const* CreatureBaseStats::GetBaseStats(uint8 level, uint8 unitClass)
 {
     return sObjectMgr.GetCreatureBaseStats(level, unitClass);
 }
